+++ hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/security/TestOzoneDelegationTokenSecretManager.java
@@ -53,8 +53,8 @@ import java.security.cert.X509Certificate;
 import java.util.HashMap;
 import java.util.Map;
 
+import static org.apache.hadoop.ozone.om.OMConfigKeys.OZONE_OM_RATIS_ENABLE_KEY;
 import static org.apache.hadoop.ozone.protocol.proto.OzoneManagerProtocolProtos.OMTokenProto.Type.S3AUTHINFO;
-
 
 /**
  * Test class for {@link OzoneDelegationTokenSecretManager}.
@@ -109,6 +109,14 @@ public class TestOzoneDelegationTokenSecretManager {
 
   private OzoneConfiguration createNewTestPath() throws IOException {
     OzoneConfiguration config = new OzoneConfiguration();
+    // When ratis is enabled, tokens are not updated to the store directly by
+    // OzoneDelegationTokenSecretManager. Tokens are updated via Ratis
+    // through the DoubleBuffer. Hence, to test
+    // OzoneDelegationTokenSecretManager, we should disable OM Ratis.
+    // TODO: Once HA and non-HA code paths are merged in
+    //  OzoneDelegationTokenSecretManager, this test should be updated to
+    //  test both ratis enabled and disabled case.
+    config.setBoolean(OZONE_OM_RATIS_ENABLE_KEY, false);
     File newFolder = folder.newFolder();
     if (!newFolder.exists()) {
       Assert.assertTrue(newFolder.mkdirs());
